requirements.txt
flask==3.0.3
python-dotenv==1.0.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0
google-generativeai==0.7.2
numpy==1.26.4
